Make config injection methods static
This patch marks the two public injection methods in the delegator
factory as static, allowing them to be used without instantiating the
factory.

It also correctly places the tests for the delegator factory, and moves
required test assests in a tree beneath it.
